Refactor materialloader test setup
